Create bind.c
Was able to write this out in 7 minutes without the aid of any book.
I now know and understand the principles of creating, addressing, and binding the socket, listening to it, and closing the socket. So far everything seems to compile and run flawlessly. 

/*
Followup:
Still need to work on concepts like open() read(), write(), etc as it pertains to the use of internet sockets.
Researching how you can listen to a socket without first "opening" it. Doesn't make sense. 

Yes, I understand that some of the includes are not actually used, but in a real program, it is not unthinkable that one would use strings, maths, and file handling when otherwise dealing with sockets.
*/